tests/wpt/web-platform-tests/html/canvas/offscreen/drawing-images-to-the-canvas/2d.drawImage.negativedest.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.drawImage.negativedest</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.drawImage.negativedest</h1>
<p class="desc">Negative destination width/height represents the correct rectangle</p>


<script>
var t = async_test("Negative destination width/height represents the correct rectangle");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(async function() {

var canvas = new OffscreenCanvas(100, 50);
var ctx = canvas.getContext('2d');

await (function(canvas, ctx) {
ctx.fillStyle = '#f00';
ctx.fillRect(0, 0, 100, 50);
fetch('/images/ggrr-256x256.png')
  .then(response => response.blob())
    .then(blob => {
      createImageBitmap(blob)
        .then(bitmap => {
        ctx.drawImage(bitmap, 100, 78, 50, 50, 0, 50, 50, -50);
        ctx.drawImage(bitmap, 100, 128, 50, -50, 100, 50, -50, -50);
        _assertPixelApprox(canvas, 1,1, 0,255,0,255, 2);
        _assertPixelApprox(canvas, 1,48, 0,255,0,255, 2);
        _assertPixelApprox(canvas, 98,1, 0,255,0,255, 2);
        _assertPixelApprox(canvas, 98,48, 0,255,0,255, 2);
        _assertPixelApprox(canvas, 48,1, 0,255,0,255, 2);
        _assertPixelApprox(canvas, 48,48, 0,255,0,255, 2);
        _assertPixelApprox(canvas, 51,1, 0,255,0,255, 2);
        _assertPixelApprox(canvas, 51,48, 0,255,0,255, 2);
        _assertPixelApprox(canvas, 25,25, 0,255,0,255, 2);
        _assertPixelApprox(canvas, 75,25, 0,255,0,255, 2);
    });
});
})(canvas, ctx);
t.done();

});
</script>
